ensres folders exist

+++ offline_data_parsing/run_preprocess_data.py
@@ -4,7 +4,9 @@ import os
 import shutil
 
 export_folder = "../src/assets/data/preprocessed_data/"
-shutil.rmtree(export_folder)
+if os.path.exists(export_folder):
+    shutil.rmtree(export_folder)
 os.makedirs(export_folder)
+
 fetch_foods.format_and_save_usda_food_data(export_folder)
 process_studies.preprocess_studies(export_folder)
